reserva_hotel/requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
debugpy==1.8.1
Django==5.0.3
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-filter==24.1
django-redis==5.4.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
freezegun==1.4.0
inflection==0.5.1
kombu==5.3.5
packaging==24.0
prompt-toolkit==3.0.43
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
